inherent InterestRateModule in LendingPool

+++ src/libraries/DataTypes.sol
@@ -1,10 +1,10 @@
 library DataTypes {
     // TODO: Optimize the variables
     struct InterestRateConfiguration {
-        uint baseRate;
-        uint lowSlope;
-        uint highSlope;
-        uint utilisationThreshold;
+        uint64 baseRate;
+        uint64 lowSlope;
+        uint64 highSlope;
+        uint64 utilisationThreshold;
     }
 
 }
